citygroves/backend/flesh/pyproject.toml
[tool.poetry]
name = "citygroves.backend"
version = "0.1.0"
description = ""
authors = ["Damian Krystkiewicz <damian.krystkiewicz@gmail.com>"]

[tool.black]
line-length = 120
target-version = ['py37']

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0"
Pillow = "^6.2"
loguru = "^0.4.0"
graypy = "^2.1"
django-click = "^2.1"
psycopg2-binary = "^2.8"
djangorestframework = "^3.11"
celery = "^4.3"
django-environ = "^0.4.5"
ipython = "^7.10"
pytest = "^5.3"
pytest-django = "^3.7"
pytest-socket = "^0.3.3"
pytest-env = "^0.6.2"
pyyaml = "^5.2"
redis = "^3.3"
pdfrw = "^0.4.0"
factory_boy = "^2.12"
pytest-factoryboy = "^2.0"
django-extensions = "^2.2"
rhei = "^0.5.2"
django-cors-headers = "^3.2"
tqdm = "^4.41"
watchdog = "^0.9.0"
django-filter = "^2.2"
pypdftk = "^0.4.0"
pyPdf2 = "^1.26"
flake8 = "^3.7"
google-api-python-client = "^1.7"
google-auth-httplib2 = "^0.0.3"
google-auth-oauthlib = "^0.4.1"
pytest-mock = "^2.0"
isort = "^4.3"
black = {version = "^18.3-alpha.0", allows-prereleases = true}
pep8-naming = "^0.9.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
